unfinished work

+++ collinear/Point.java
@@ -62,7 +62,7 @@ public class Point implements Comparable<Point> {
     	if ((this.y == that.y) && (this.x == that.x)) return Double.NEGATIVE_INFINITY;
     	if (this.y == that.y) return +0.0;
     	if (this.x == that.x) return Double.POSITIVE_INFINITY;
-    	return (that.y - this.y)/(that.x - this.x);
+    	return (that.y - this.y)/(double)(that.x - this.x);
     }
 
     /**
